Configuration Drift Detector and Reconciler/requirements.txt
PyYAML
argparse
GitPython
Jinja2
croniter
logging